delete user

+++ handler/user.go
@@ -3,7 +3,9 @@ package handler
 import (
 	"api-fiber-gorm/database"
 	"api-fiber-gorm/model"
+	"strconv"
 
+	"github.com/dgrijalva/jwt-go"
 	"github.com/gofiber/fiber"
 	"golang.org/x/crypto/bcrypt"
 )
@@ -49,15 +51,43 @@ func CreateUser(c *fiber.Ctx) {
 
 // DeleteUser delete user
 func DeleteUser(c *fiber.Ctx) {
-	id := c.Params("id")
-	db := database.DB
-
-	var user model.User
-	db.First(&user, id)
-	if user.Username == "" {
-		c.Status(404).JSON(fiber.Map{"status": "error", "message": "No user found with ID", "data": nil})
+	type PasswordInput struct {
+		Password string `json:"password"`
+	}
+	var pi PasswordInput
+	if err := c.BodyParser(&pi); err != nil {
+		c.Status(500).JSON(fiber.Map{"status": "error", "message": "Review your input", "data": err})
 		return
 	}
+	id := c.Params("id")
+	n, err := strconv.Atoi(id)
+	if err != nil {
+		c.Status(500).JSON(fiber.Map{"status": "error", "message": "Couldn'n read id from params", "data": err})
+		return
+	}
+
+	token := c.Locals("user").(*jwt.Token)
+	claims := token.Claims.(jwt.MapClaims)
+	uid := int(claims["user_id"].(float64))
+	db := database.DB
+	var user model.User
+
+	if uid != n {
+		c.Status(fiber.StatusUnauthorized).JSON(fiber.Map{"status": "error", "message": "Couldn't delete user", "data": nil})
+		return
+	}
+
+	db.First(&user, id)
+	if user.Username == "" {
+		c.Status(fiber.StatusNotFound).JSON(fiber.Map{"status": "error", "message": "No user found with ID", "data": nil})
+		return
+	}
+
+	if !CheckPasswordHash(pi.Password, user.Password) {
+		c.Status(fiber.StatusUnauthorized).JSON(fiber.Map{"status": "error", "message": "Invalid password", "data": nil})
+		return
+	}
+
 	db.Delete(&user)
 	c.JSON(fiber.Map{"status": "success", "message": "User successfully deleted", "data": nil})
 }
